Modificato db autoincrement Modificati alcuni dao e creati EventoDao e PosizioneDao Commit di emergenza per mancanza corrente, non completato

+++ workspace/LetsMeetWeb/src/it/unisa/studenti/letsmeet/manager/Dao.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface Dao<T> {
 	
-	T get(long id) throws DaoException;
+	T get(int id) throws DaoException;
 	
 	List<T> getAll() throws DaoException;
